+++ apm-sniffer/optional-reporter-plugins/kafka-reporter-plugin/src/main/java/org/apache/skywalking/apm/agent/core/kafka/KafkaReporterPluginConfig.java
@@ -42,9 +42,7 @@ public class KafkaReporterPluginConfig {
 
             public static String TOPIC_METER = "skywalking-meters";
 
-            public static String MM_TO_SOURCE_ALIAS = "";
-
-            public static String MM_TO_SOURCE_SEPARATOR = "";
+            public static String NAMESPACE = "";
 
             public static Map<String, String> PRODUCER_CONFIG = new HashMap<>();