+++ CHANGES.md
@@ -12,7 +12,7 @@ Release Notes.
 * Improve `GlobalIdGenerator` performance.
 * Add an agent plugin to support elasticsearch7.
 * Add `jsonrpc4j` agent plugin.
-* new options to support multi skywalking cluster use same kafka cluster(plugin.kafka.mm_to_source_alias,plugin.kafka.mm_to_source_separator)
+* new options to support multi skywalking cluster use same kafka cluster(plugin.kafka.namespace)
 * resolve agent has no retries if connect kafka cluster failed when bootstrap
 
 #### OAP-Backend